rename feature.rs to line_editor.rs
